src/dev-requirements.txt
# Local Environment
python-dotenv~=1.0.1
pre-commit~=4.1.0
detect-secrets~=1.5.0

# Pinecone
pinecone~=5.4.0
pinecone-plugin-assistant~=0.3.0

# Slack
slack-bolt~=1.21.2
slack-cli-hooks~=0.0.3

# AWS Lambda
python-lambda

# Tinybird
tinybird-cli